WIP: need to move help stuff
I need to make `warg.New` take the HelpMappings as a param, since
I can't import all the help stuff or get the circular import.

I don't like this because that param rarely needs to change, but it
unblocks me, and maybe I can add a layer later to fix it.